add unit tests for client server message calls.

+++ RpcPipes.Models/PipeModels.cs
@@ -2,6 +2,6 @@ using RpcPipes.Transport;
 
 namespace RpcPipes.Models;
 
-public record RequestMessage(string Message, int DelaySeconds);
+public record RequestMessage(string Message, double DelaySeconds);
 public record ReplyMessage(string Reply);
 public record ProgressMessage(double Progress, string Stage) : IPipeProgress;
